fix(cloud): switch default cloud from SmartHome to Nethome Plus (#360)
switch default cloud from Smarthome to Nethome plus.
as Midea closed the api service in Smarthome cloud.

we just receive this temp solution from msmart-ng repo, the origin link
is :
https://github.com/mill1000/midea-msmart/issues/193
https://github.com/mill1000/midea-msmart/pull/194

anyway, it's a backup solution, and midea also will close this service
in future, so we still need to migrate to new v2 cloud api in future.