feat(a1): allow customization of default modes and speeds for dehumidifiers (#389)
This allows one to set custom values for the default **modes** and
**speeds** of a dehumidifier. The code was largely copied from other
device classes with the `set_customize` method and modified to fit this
use case. I tested it without issue on my device.

For my use case, I was trying to modify the **modes** and **speeds** in
HA to match those of the Midea 50 pt. Cube Dehumidifier with Built-in
Pump (MAD50PS1QWT). However, this is applicable to any other
dehumidifier that has different **modes** and **speeds** than the
defaults.

For the MAD50PS1QWT device, the valid custom configuration is:

`{ "speeds": {"40": "Low", "80": "High"}, "modes": {"1": "Manual", "2":
"Continuous", "4": "Max"} }`

Generally, this custom configuration should be the same for all Midea
Cube Dehumidifiers and should solve the problems discussed in [Issue
570](https://github.com/wuwentao/midea_ac_lan/issues/570).

Co-authored-by: Albert Gorski <sigmundgorski@gmail.com>